rpicamera: support two CSI cameras at once (#1573)
