+++ internal/conf/conf_test.go
@@ -202,7 +202,7 @@ func TestConfErrors(t *testing.T) {
 				"    source: rpiCamera\n" +
 				"  cam2:\n" +
 				"    source: rpiCamera\n",
-			"'rpiCamera' is used as source in two paths ('cam1' and 'cam2')",
+			"'rpiCamera' with same camera ID 0 is used as source in two paths, 'cam1' and 'cam2'",
 		},
 	} {
 		t.Run(ca.name, func(t *testing.T) {
